backend/requirements.txt
Flask
firebase-admin
flask-cors
python-dotenv
